added-and-changed-files
added files in util
added a file in middleware
changed the code for the authRouter.js so it can wrap the functions in the express router and using the router.post() and router.get() methods, ensuring compatibility with server.js setup.
added userRoutes.js defines the API endpoints responsible for User Profile Management and Administrator User Management.